Rename README.md to Baseline
Created as a place holder for Scyion's Space in Altspace.
An Universe VR World and Advanced  Unity Development Project. 
Opening project in fact. 
The first version was based on a 'plane' that we created in Unity and imported via the Altspace Template.
The second version was based on a 'terrain' versus a 'plane'. 
I have been editing that version since including modifying the terrain and adding a beach / coast area to our current terrain.